chore: refactor functions out of index.ts into dedicated files (file utils, asar utils, integrity generator)
